--- src/main/java/tech/pragmat/CountryRegionService/repository/CountryRegionRepository.java
@@ -1,17 +1,0 @@
-package tech.pragmat.CountryRegionService.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import tech.pragmat.CountryRegionService.model.entity.CountryRegion;
-
-import java.util.List;
-
-@Repository
-public interface CountryRegionRepository extends CrudRepository<CountryRegion, Integer> {
-
-    List<CountryRegion> findAll();
-
-    CountryRegion findFirstByCountry(String name);
-
-    CountryRegion deleteByCountry(String name);
-}
