+++ src/main/java/tech/pragmat/countryregionservice/feign/CountryRegionClient.java
@@ -1,10 +1,10 @@
-package tech.pragmat.CountryRegionService.feign;
+package tech.pragmat.countryregionservice.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(name = "requestToCountryRegionService", url = "http://localhost:9090/Geo2LiteCountry")
+@FeignClient(name = "requestToCountryRegionService", url = "${cr.url}")
 public interface CountryRegionClient {
 
     @GetMapping("/getCountryByIp")
