+++ src/main/java/tech/pragmat/countryregionservice/feign/GeoNameClient.java
@@ -1,9 +1,9 @@
-package tech.pragmat.CountryRegionService.feign;
+package tech.pragmat.countryregionservice.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 
-@FeignClient(name = "requestToGeoNameService", url = "https://download.geonames.org/")
+@FeignClient(name = "requestToGeoNameService", url = "${geo.url}")
 public interface GeoNameClient {
 
     @GetMapping("export/dump/countryInfo.txt")
