+++ src/main/java/tech/pragmat/countryregionservice/model/CountryRegion.java
@@ -1,6 +1,7 @@
-package tech.pragmat.CountryRegionService.model.entity;
+package tech.pragmat.countryregionservice.model;
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -9,6 +10,7 @@ import javax.persistence.Id;
 
 @Data
 @Entity(name = "country_region")
+@NoArgsConstructor
 public class CountryRegion {
 
     @Id
@@ -23,7 +25,4 @@ public class CountryRegion {
         this.region = region;
     }
 
-    public CountryRegion() {
-
-    }
 }
