+++ src/main/java/tech/pragmat/countryregionservice/web/controller/ConnectionWithGeoLite2CountryController.java
@@ -1,18 +1,18 @@
-package tech.pragmat.CountryRegionService.web.controller;
+package tech.pragmat.countryregionservice.web.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import tech.pragmat.CountryRegionService.service.CountyRegionFeignService;
+import tech.pragmat.countryregionservice.service.CountyRegionFeignService;
 
 @Slf4j
 @RestController
 @RequestMapping("/connectGeo2LiteService")
 public class ConnectionWithGeoLite2CountryController {
 
-    private final CountyRegionFeignService countyRegionFeignService;
+    private final   CountyRegionFeignService countyRegionFeignService;
 
     public ConnectionWithGeoLite2CountryController(CountyRegionFeignService countyRegionFeignService) {
         this.countyRegionFeignService = countyRegionFeignService;
